re: #2902 change wp_id to string and update default sorting for query results
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@7996 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/java/org.wcs.smart.patrol/src/org/wcs/smart/patrol/internal/ui/editor/WaypointSorter.java
@@ -90,7 +90,7 @@ public class WaypointSorter extends ViewerComparator{
 	
 	private int compareValue(Waypoint wp1, Waypoint wp2){
 		if (sortColumn == OtColumn.ID){
-			return ((Integer)wp1.getId()).compareTo(((Integer)wp2.getId()));
+			return wp1.getId().compareTo(wp2.getId());
 		}else if (sortColumn == OtColumn.TIME){
 			return wp1.getDateTime().compareTo(wp2.getDateTime());
 			//return ((Integer)wp1.getTime()).compareTo(((Integer)wp2.getTime()));
